+++ packages/router/__tests__/router-test.ts
@@ -14,7 +14,7 @@ import type {
 import {
   createMemoryHistory,
   createRouter,
-  unstable_createStaticHandler as createStaticHandler,
+  createStaticHandler,
   defer,
   ErrorResponse,
   IDLE_FETCHER,
@@ -875,7 +875,57 @@ afterEach(() => {
 
 describe("a router", () => {
   describe("init", () => {
-    it("with initial values", async () => {
+    it("initial state w/o hydrationData", async () => {
+      let history = createMemoryHistory({ initialEntries: ["/"] });
+      let router = createRouter({
+        routes: [
+          {
+            id: "root",
+            path: "/",
+            hasErrorBoundary: true,
+            loader: () => Promise.resolve(),
+          },
+        ],
+        history,
+      });
+      expect(router.state).toEqual({
+        historyAction: "POP",
+        loaderData: {},
+        actionData: null,
+        errors: null,
+        location: {
+          hash: "",
+          key: expect.any(String),
+          pathname: "/",
+          search: "",
+          state: null,
+        },
+        matches: [
+          {
+            params: {},
+            pathname: "/",
+            pathnameBase: "/",
+            route: {
+              hasErrorBoundary: true,
+              id: "root",
+              loader: expect.any(Function),
+              path: "/",
+            },
+          },
+        ],
+        initialized: false,
+        navigation: {
+          location: undefined,
+          state: "idle",
+        },
+        preventScrollReset: false,
+        restoreScrollPosition: null,
+        revalidation: "idle",
+        fetchers: new Map(),
+      });
+    });
+
+    it("initial state w/hydrationData values", async () => {
       let history = createMemoryHistory({ initialEntries: ["/"] });
       let router = createRouter({
         routes: [
@@ -930,7 +980,7 @@ describe("a router", () => {
           state: "idle",
         },
         preventScrollReset: false,
-        restoreScrollPosition: null,
+        restoreScrollPosition: false,
         revalidation: "idle",
         fetchers: new Map(),
       });
@@ -1602,7 +1652,7 @@ describe("a router", () => {
         },
         nextUrl: new URL("http://localhost/params/aValue/bValue"),
         defaultShouldRevalidate: false,
-        actionResult: null,
+        actionResult: undefined,
       });
       rootLoader.mockClear();
       shouldRevalidate.mockClear();
@@ -1617,6 +1667,112 @@ describe("a router", () => {
       });
       await tick();
       expect(rootLoader.mock.calls.length).toBe(0);
+
+      router.dispose();
+    });
+
+    it("includes submission on actions that return data", async () => {
+      let shouldRevalidate = jest.fn(() => true);
+
+      let history = createMemoryHistory({ initialEntries: ["/child"] });
+      let router = createRouter({
+        history,
+        routes: [
+          {
+            path: "/",
+            id: "root",
+            loader: () => "ROOT",
+            shouldRevalidate,
+            children: [
+              {
+                path: "child",
+                id: "child",
+                loader: () => "CHILD",
+                action: () => "ACTION",
+              },
+            ],
+          },
+        ],
+      });
+      router.initialize();
+
+      // Initial load - no existing data, should always call loader and should
+      // not give use ability to opt-out
+      await tick();
+      router.navigate("/child", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      await tick();
+      expect(shouldRevalidate.mock.calls.length).toBe(1);
+      // @ts-expect-error
+      let arg = shouldRevalidate.mock.calls[0][0];
+      expect(arg).toMatchObject({
+        currentParams: {},
+        currentUrl: new URL("http://localhost/child"),
+        nextParams: {},
+        nextUrl: new URL("http://localhost/child"),
+        defaultShouldRevalidate: true,
+        formMethod: "post",
+        formAction: "/child",
+        formEncType: "application/x-www-form-urlencoded",
+        actionResult: "ACTION",
+      });
+      // @ts-expect-error
+      expect(Object.fromEntries(arg.formData)).toEqual({ key: "value" });
+
+      router.dispose();
+    });
+
+    it("includes submission on actions that return redirects", async () => {
+      let shouldRevalidate = jest.fn(() => true);
+
+      let history = createMemoryHistory({ initialEntries: ["/child"] });
+      let router = createRouter({
+        history,
+        routes: [
+          {
+            path: "/",
+            id: "root",
+            loader: () => "ROOT",
+            shouldRevalidate,
+            children: [
+              {
+                path: "child",
+                id: "child",
+                loader: () => "CHILD",
+                action: () => redirect("/"),
+              },
+            ],
+          },
+        ],
+      });
+      router.initialize();
+
+      // Initial load - no existing data, should always call loader and should
+      // not give use ability to opt-out
+      await tick();
+      router.navigate("/child", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      await tick();
+      expect(shouldRevalidate.mock.calls.length).toBe(1);
+      // @ts-expect-error
+      let arg = shouldRevalidate.mock.calls[0][0];
+      expect(arg).toMatchObject({
+        currentParams: {},
+        currentUrl: new URL("http://localhost/child"),
+        nextParams: {},
+        nextUrl: new URL("http://localhost/"),
+        defaultShouldRevalidate: true,
+        formMethod: "post",
+        formAction: "/child",
+        formEncType: "application/x-www-form-urlencoded",
+        actionResult: undefined,
+      });
+      // @ts-expect-error
+      expect(Object.fromEntries(arg.formData)).toEqual({ key: "value" });
 
       router.dispose();
     });
@@ -1844,7 +2000,8 @@ describe("a router", () => {
         data: "FETCH",
       });
 
-      expect(shouldRevalidate.mock.calls[0][0]).toMatchInlineSnapshot(`
+      let arg = shouldRevalidate.mock.calls[0][0];
+      expect(arg).toMatchInlineSnapshot(`
         Object {
           "actionResult": "FETCH",
           "currentParams": Object {},
@@ -1854,6 +2011,63 @@ describe("a router", () => {
           "formData": FormData {},
           "formEncType": "application/x-www-form-urlencoded",
           "formMethod": "post",
+          "nextParams": Object {},
+          "nextUrl": "http://localhost/",
+        }
+      `);
+      expect(Object.fromEntries(arg.formData)).toEqual({ key: "value" });
+
+      router.dispose();
+    });
+
+    it("applies to fetcher submissions when action redirects", async () => {
+      let shouldRevalidate = jest.fn((args) => true);
+
+      let history = createMemoryHistory();
+      let router = createRouter({
+        history,
+        routes: [
+          {
+            path: "",
+            id: "root",
+
+            children: [
+              {
+                path: "/",
+                id: "index",
+                loader: () => "INDEX",
+                shouldRevalidate,
+              },
+              {
+                path: "/fetch",
+                id: "fetch",
+                action: () => redirect("/"),
+              },
+            ],
+          },
+        ],
+      });
+      router.initialize();
+      await tick();
+
+      let key = "key";
+      router.fetch(key, "root", "/fetch", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      await tick();
+      expect(router.state.fetchers.get(key)).toMatchObject({
+        state: "idle",
+        data: undefined,
+      });
+
+      let arg = shouldRevalidate.mock.calls[0][0];
+      expect(arg).toMatchInlineSnapshot(`
+        Object {
+          "actionResult": undefined,
+          "currentParams": Object {},
+          "currentUrl": "http://localhost/",
+          "defaultShouldRevalidate": true,
           "nextParams": Object {},
           "nextUrl": "http://localhost/",
         }
@@ -2256,6 +2470,44 @@ describe("a router", () => {
           child: new Error("Kaboom!"),
         });
       });
+
+      it("clears previous loaderData at that route", async () => {
+        let t = setup({
+          routes: [
+            {
+              path: "/",
+              id: "parent",
+              loader: true,
+              children: [
+                {
+                  path: "/child",
+                  id: "child",
+                  hasErrorBoundary: true,
+                  loader: true,
+                },
+              ],
+            },
+          ],
+        });
+        let nav = await t.navigate("/child");
+        await nav.loaders.parent.resolve("PARENT");
+        await nav.loaders.child.resolve("CHILD");
+        expect(t.router.state.loaderData).toEqual({
+          parent: "PARENT",
+          child: "CHILD",
+        });
+        expect(t.router.state.errors).toEqual(null);
+
+        let nav2 = await t.navigate("/child");
+        await nav2.loaders.parent.resolve("PARENT2");
+        await nav2.loaders.child.reject(new Error("Kaboom!"));
+        expect(t.router.state.loaderData).toEqual({
+          parent: "PARENT2",
+        });
+        expect(t.router.state.errors).toEqual({
+          child: new Error("Kaboom!"),
+        });
+      });
     });
 
     describe("with an error boundary above the throwing route", () => {
@@ -2469,6 +2721,35 @@ describe("a router", () => {
       expect(t.router.state.location.pathname).toEqual("/foo");
     });
 
+    it("navigates correctly using POP navigations across actions to new locations", async () => {
+      let t = initializeTmTest();
+
+      // Navigate to /foo
+      let A = await t.navigate("/foo");
+      await A.loaders.foo.resolve("FOO");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+
+      // Navigate to /bar
+      let B = await t.navigate("/bar");
+      await B.loaders.bar.resolve("BAR");
+      expect(t.router.state.location.pathname).toEqual("/bar");
+
+      // Post to /baz (should not replace)
+      let C = await t.navigate("/baz", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      await C.actions.baz.resolve("BAZ ACTION");
+      await C.loaders.root.resolve("ROOT");
+      await C.loaders.baz.resolve("BAZ");
+      expect(t.router.state.location.pathname).toEqual("/baz");
+
+      // POP to /bar
+      let D = await t.navigate(-1);
+      await D.loaders.bar.resolve("BAR");
+      expect(t.router.state.location.pathname).toEqual("/bar");
+    });
+
     it("navigates correctly using POP navigations across action errors", async () => {
       let t = initializeTmTest();
 
@@ -2585,6 +2866,42 @@ describe("a router", () => {
       expect(t.router.state.location.key).not.toBe(postBarKey);
     });
 
+    it("navigates correctly using POP navigations across action redirects to the same location", async () => {
+      let t = initializeTmTest();
+
+      // Navigate to /foo
+      let A = await t.navigate("/foo");
+      let fooKey = t.router.state.navigation.location?.key;
+      await A.loaders.foo.resolve("FOO");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+
+      // Navigate to /bar
+      let B = await t.navigate("/bar");
+      await B.loaders.bar.resolve("BAR");
+      expect(t.router.state.historyAction).toEqual("PUSH");
+      expect(t.router.state.location.pathname).toEqual("/bar");
+
+      // Post to /bar, redirect to /bar
+      let C = await t.navigate("/bar", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      let postBarKey = t.router.state.navigation.location?.key;
+      let D = await C.actions.bar.redirect("/bar");
+      await D.loaders.root.resolve("ROOT");
+      await D.loaders.bar.resolve("BAR");
+      expect(t.router.state.historyAction).toEqual("REPLACE");
+      expect(t.router.state.location.pathname).toEqual("/bar");
+
+      // POP to /foo
+      let E = await t.navigate(-1);
+      await E.loaders.foo.resolve("FOO");
+      expect(t.router.state.historyAction).toEqual("POP");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+      expect(t.router.state.location.key).toBe(fooKey);
+      expect(t.router.state.location.key).not.toBe(postBarKey);
+    });
+
     it("navigates correctly using POP navigations across <Form replace> redirects", async () => {
       let t = initializeTmTest();
 
@@ -2617,6 +2934,67 @@ describe("a router", () => {
       expect(t.router.state.historyAction).toEqual("POP");
       expect(t.router.state.location.pathname).toEqual("/foo");
     });
+
+    it("should respect explicit replace:false on non-redirected actions to new locations", async () => {
+      // start at / (history stack: [/])
+      let t = initializeTmTest();
+
+      // Link to /foo (history stack: [/, /foo])
+      let A = await t.navigate("/foo");
+      await A.loaders.root.resolve("ROOT");
+      await A.loaders.foo.resolve("FOO");
+      expect(t.router.state.historyAction).toEqual("PUSH");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+
+      // POST /bar (history stack: [/, /foo, /bar])
+      let B = await t.navigate("/bar", {
+        formMethod: "post",
+        formData: createFormData({ gosh: "dang" }),
+        replace: false,
+      });
+      await B.actions.bar.resolve("BAR");
+      await B.loaders.root.resolve("ROOT");
+      await B.loaders.bar.resolve("BAR");
+      expect(t.router.state.historyAction).toEqual("PUSH");
+      expect(t.router.state.location.pathname).toEqual("/bar");
+
+      // POP /foo (history stack: [GET /, GET /foo])
+      let C = await t.navigate(-1);
+      await C.loaders.foo.resolve("FOO");
+      expect(t.router.state.historyAction).toEqual("POP");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+    });
+
+    it("should respect explicit replace:false on non-redirected actions to the same location", async () => {
+      // start at / (history stack: [/])
+      let t = initializeTmTest();
+
+      // Link to /foo (history stack: [/, /foo])
+      let A = await t.navigate("/foo");
+      await A.loaders.root.resolve("ROOT");
+      await A.loaders.foo.resolve("FOO");
+      expect(t.router.state.historyAction).toEqual("PUSH");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+
+      // POST /foo (history stack: [/, /foo, /foo])
+      let B = await t.navigate("/foo", {
+        formMethod: "post",
+        formData: createFormData({ gosh: "dang" }),
+        replace: false,
+      });
+      await B.actions.foo.resolve("FOO2 ACTION");
+      await B.loaders.root.resolve("ROOT2");
+      await B.loaders.foo.resolve("FOO2");
+      expect(t.router.state.historyAction).toEqual("PUSH");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+
+      // POP /foo (history stack: [/, /foo])
+      let C = await t.navigate(-1);
+      await C.loaders.root.resolve("ROOT3");
+      await C.loaders.foo.resolve("FOO3");
+      expect(t.router.state.historyAction).toEqual("POP");
+      expect(t.router.state.location.pathname).toEqual("/foo");
+    });
   });
 
   describe("submission navigations", () => {
@@ -2630,6 +3008,9 @@ describe("a router", () => {
 
       await A.actions.foo.resolve("FOO ACTION");
       expect(A.loaders.root.stub.mock.calls.length).toBe(1);
+      expect(t.router.state.actionData).toEqual({
+        foo: "FOO ACTION",
+      });
 
       let B = await A.loaders.foo.redirect("/bar");
       await A.loaders.root.reject("ROOT ERROR");
@@ -2637,6 +3018,7 @@ describe("a router", () => {
       await B.loaders.bar.resolve("BAR LOADER");
       expect(B.loaders.root.stub.mock.calls.length).toBe(1);
       expect(t.router.state).toMatchObject({
+        actionData: null,
         loaderData: {
           root: "ROOT LOADER 2",
           bar: "BAR LOADER",
@@ -2668,8 +3050,11 @@ describe("a router", () => {
       });
       expect(A.loaders.root.stub.mock.calls.length).toBe(0);
 
-      await A.actions.foo.resolve(null);
+      await A.actions.foo.resolve("FOO ACTION");
       expect(A.loaders.root.stub.mock.calls.length).toBe(1);
+      expect(t.router.state.actionData).toEqual({
+        foo: "FOO ACTION",
+      });
 
       await A.loaders.foo.resolve("A LOADER");
       expect(t.router.state.navigation.state).toBe("loading");
@@ -2680,6 +3065,9 @@ describe("a router", () => {
 
       await A.loaders.root.resolve("ROOT LOADER");
       expect(t.router.state.navigation.state).toBe("idle");
+      expect(t.router.state.actionData).toEqual({
+        foo: "FOO ACTION", // kept around on action reload
+      });
       expect(t.router.state.loaderData).toEqual({
         foo: "A LOADER",
         root: "ROOT LOADER",
@@ -2852,6 +3240,45 @@ describe("a router", () => {
       expect(t.router.state.actionData).toBeNull();
       expect(t.router.state.loaderData).toEqual({
         other: "OTHER",
+      });
+    });
+
+    it("removes action data after action redirect to current location", async () => {
+      let t = setup({
+        routes: [
+          {
+            path: "/",
+            id: "index",
+            action: true,
+            loader: true,
+          },
+        ],
+      });
+      let A = await t.navigate("/", {
+        formMethod: "post",
+        formData: createFormData({ gosh: "" }),
+      });
+      await A.actions.index.resolve({ error: "invalid" });
+      expect(t.router.state.actionData).toEqual({
+        index: { error: "invalid" },
+      });
+
+      let B = await t.navigate("/", {
+        formMethod: "post",
+        formData: createFormData({ gosh: "dang" }),
+      });
+
+      let C = await B.actions.index.redirectReturn("/");
+      expect(t.router.state.actionData).toEqual({
+        index: { error: "invalid" },
+      });
+      expect(t.router.state.loaderData).toEqual({});
+
+      await C.loaders.index.resolve("NEW");
+
+      expect(t.router.state.actionData).toBeNull();
+      expect(t.router.state.loaderData).toEqual({
+        index: "NEW",
       });
     });
 
@@ -3236,6 +3663,99 @@ describe("a router", () => {
         });
       });
     });
+
+    it("clears previous actionData at the throwing route", async () => {
+      let t = setup({
+        routes: [
+          {
+            path: "/",
+            id: "parent",
+            loader: true,
+            children: [
+              {
+                path: "/child",
+                id: "child",
+                hasErrorBoundary: true,
+                action: true,
+                loader: true,
+              },
+            ],
+          },
+        ],
+      });
+      let nav = await t.navigate("/child", {
+        formMethod: "post",
+        formData: createFormData({ key: "value" }),
+      });
+      await nav.actions.child.resolve("ACTION");
+      await nav.loaders.parent.resolve("PARENT");
+      await nav.loaders.child.resolve("CHILD");
+      expect(t.router.state.actionData).toEqual({
+        child: "ACTION",
+      });
+      expect(t.router.state.loaderData).toEqual({
+        parent: "PARENT",
+        child: "CHILD",
+      });
+      expect(t.router.state.errors).toEqual(null);
+
+      let nav2 = await t.navigate("/child", {
+        formMethod: "post",
+        formData: createFormData({ key2: "value2" }),
+      });
+      await nav2.actions.child.reject(new Error("Kaboom!"));
+      await nav2.loaders.parent.resolve("PARENT2");
+      expect(t.router.state.actionData).toEqual(null);
+      expect(t.router.state.loaderData).toEqual({
+        parent: "PARENT2",
+      });
+      expect(t.router.state.errors).toEqual({
+        child: new Error("Kaboom!"),
+      });
+    });
+
+    it("does not clear previous loaderData at the handling route", async () => {
+      let t = setup({
+        routes: [
+          {
+            path: "/",
+            id: "parent",
+            loader: true,
+            hasErrorBoundary: true,
+            children: [
+              {
+                path: "/child",
+                id: "child",
+                action: true,
+                loader: true,
+              },
+            ],
+          },
+        ],
+      });
+      let nav = await t.navigate("/child");
+      await nav.loaders.parent.resolve("PARENT");
+      await nav.loaders.child.resolve("CHILD");
+      expect(t.router.state.actionData).toEqual(null);
+      expect(t.router.state.loaderData).toEqual({
+        parent: "PARENT",
+        child: "CHILD",
+      });
+      expect(t.router.state.errors).toEqual(null);
+
+      let nav2 = await t.navigate("/child", {
+        formMethod: "post",
+        formData: createFormData({ key2: "value2" }),
+      });
+      await nav2.actions.child.reject(new Error("Kaboom!"));
+      expect(t.router.state.actionData).toEqual(null);
+      expect(t.router.state.loaderData).toEqual({
+        parent: "PARENT",
+      });
+      expect(t.router.state.errors).toEqual({
+        parent: new Error("Kaboom!"),
+      });
+    });
   });
 
   describe("navigation states", () => {
@@ -3372,9 +3892,9 @@ describe("a router", () => {
       });
       let navigation = t.router.state.navigation;
       expect(navigation.state).toBe("loading");
-      expect(navigation.formData).toBeUndefined();
-      expect(navigation.formMethod).toBeUndefined();
-      expect(navigation.formEncType).toBeUndefined();
+      expect(navigation.formData).toEqual(createFormData({ gosh: "dang" }));
+      expect(navigation.formMethod).toBe("get");
+      expect(navigation.formEncType).toBe("application/x-www-form-urlencoded");
       expect(navigation.location).toMatchObject({
         pathname: "/foo",
         search: "?gosh=dang",
@@ -3400,9 +3920,9 @@ describe("a router", () => {
 
       let navigation = t.router.state.navigation;
       expect(navigation.state).toBe("loading");
-      expect(navigation.formData).toBeUndefined();
-      expect(navigation.formMethod).toBeUndefined();
-      expect(navigation.formEncType).toBeUndefined();
+      expect(navigation.formData).toEqual(createFormData({ gosh: "dang" }));
+      expect(navigation.formMethod).toBe("get");
+      expect(navigation.formEncType).toBe("application/x-www-form-urlencoded");
       expect(navigation.location?.pathname).toBe("/bar");
 
       await B.loaders.bar.resolve("B");
@@ -3872,8 +4392,10 @@ describe("a router", () => {
         pathname: "/tasks",
         search: "?key=value",
       });
-      expect(t.router.state.navigation.formMethod).toBeUndefined();
-      expect(t.router.state.navigation.formData).toBeUndefined();
+      expect(t.router.state.navigation.formMethod).toBe("get");
+      expect(t.router.state.navigation.formData).toEqual(
+        createFormData({ key: "value" })
+      );
     });
 
     it("converts formData to URLSearchParams for formMethod=get", async () => {
@@ -3890,8 +4412,10 @@ describe("a router", () => {
         pathname: "/tasks",
         search: "?key=value",
       });
-      expect(t.router.state.navigation.formMethod).toBeUndefined();
-      expect(t.router.state.navigation.formData).toBeUndefined();
+      expect(t.router.state.navigation.formMethod).toBe("get");
+      expect(t.router.state.navigation.formData).toEqual(
+        createFormData({ key: "value" })
+      );
     });
 
     it("does not preserve existing 'action' URLSearchParams for formMethod='get'", async () => {
@@ -3908,8 +4432,10 @@ describe("a router", () => {
         pathname: "/tasks",
         search: "?key=2",
       });
-      expect(t.router.state.navigation.formMethod).toBeUndefined();
-      expect(t.router.state.navigation.formData).toBeUndefined();
+      expect(t.router.state.navigation.formMethod).toBe("get");
+      expect(t.router.state.navigation.formData).toEqual(
+        createFormData({ key: "2" })
+      );
     });
 
     it("preserves existing 'action' URLSearchParams for formMethod='post'", async () => {
@@ -4656,6 +5182,20 @@ describe("a router", () => {
           signal: nav3.loaders.tasks.stub.mock.calls[0][0].request.signal,
         }),
       });
+
+      let nav4 = await t.navigate("/tasks#hash", {
+        formData: createFormData({ foo: "bar" }),
+      });
+      expect(nav4.loaders.tasks.stub).toHaveBeenCalledWith({
+        params: {},
+        request: new Request("http://localhost/tasks?foo=bar", {
+          signal: nav4.loaders.tasks.stub.mock.calls[0][0].request.signal,
+        }),
+      });
+
+      expect(t.router.state.navigation.formAction).toBe("/tasks");
+      expect(t.router.state.navigation?.location?.pathname).toBe("/tasks");
+      expect(t.router.state.navigation?.location?.search).toBe("?foo=bar");
     });
 
     it("handles errors thrown from loaders", async () => {
@@ -5815,19 +6355,77 @@ describe("a router", () => {
   });
 
   describe("scroll restoration", () => {
-    it("restores scroll on navigations", async () => {
+    // Version of TASK_ROUTES with no root loader to allow for initialized
+    // hydrationData:null usage
+    const SCROLL_ROUTES: TestRouteObject[] = [
+      {
+        path: "/",
+        children: [
+          {
+            id: "index",
+            index: true,
+            loader: true,
+          },
+          {
+            id: "tasks",
+            path: "tasks",
+            loader: true,
+            action: true,
+          },
+          {
+            path: "no-loader",
+          },
+        ],
+      },
+    ];
+
+    it("restores scroll on initial load (w/o hydrationData)", async () => {
       let t = setup({
-        routes: TASK_ROUTES,
+        routes: SCROLL_ROUTES,
+        initialEntries: ["/no-loader"],
+      });
+
+      expect(t.router.state.restoreScrollPosition).toBe(null);
+      expect(t.router.state.preventScrollReset).toBe(false);
+
+      // Assume initial location had a saved position
+      let positions = { default: 50 };
+      t.router.enableScrollRestoration(positions, () => 0);
+      expect(t.router.state.restoreScrollPosition).toBe(50);
+    });
+
+    it("restores scroll on initial load (w/ hydrationData)", async () => {
+      let t = setup({
+        routes: SCROLL_ROUTES,
         initialEntries: ["/"],
         hydrationData: {
           loaderData: {
-            root: "ROOT_DATA",
+            index: "INDEX",
+          },
+        },
+      });
+
+      expect(t.router.state.restoreScrollPosition).toBe(false);
+      expect(t.router.state.preventScrollReset).toBe(false);
+
+      // Assume initial location had a saved position
+      let positions = { default: 50 };
+      t.router.enableScrollRestoration(positions, () => 0);
+      expect(t.router.state.restoreScrollPosition).toBe(false);
+    });
+
+    it("restores scroll on navigations", async () => {
+      let t = setup({
+        routes: SCROLL_ROUTES,
+        initialEntries: ["/"],
+        hydrationData: {
+          loaderData: {
             index: "INDEX_DATA",
           },
         },
       });
 
-      expect(t.router.state.restoreScrollPosition).toBe(null);
+      expect(t.router.state.restoreScrollPosition).toBe(false);
       expect(t.router.state.preventScrollReset).toBe(false);
 
       let positions = {};
@@ -5861,17 +6459,16 @@ describe("a router", () => {
 
     it("restores scroll using custom key", async () => {
       let t = setup({
-        routes: TASK_ROUTES,
+        routes: SCROLL_ROUTES,
         initialEntries: ["/"],
         hydrationData: {
           loaderData: {
-            root: "ROOT_DATA",
             index: "INDEX_DATA",
           },
         },
       });
 
-      expect(t.router.state.restoreScrollPosition).toBe(null);
+      expect(t.router.state.restoreScrollPosition).toBe(false);
       expect(t.router.state.preventScrollReset).toBe(false);
 
       let positions = { "/tasks": 100 };
@@ -5890,17 +6487,16 @@ describe("a router", () => {
 
     it("does not restore scroll on submissions", async () => {
       let t = setup({
-        routes: TASK_ROUTES,
+        routes: SCROLL_ROUTES,
         initialEntries: ["/"],
         hydrationData: {
           loaderData: {
-            root: "ROOT_DATA",
             index: "INDEX_DATA",
           },
         },
       });
 
-      expect(t.router.state.restoreScrollPosition).toBe(null);
+      expect(t.router.state.restoreScrollPosition).toBe(false);
       expect(t.router.state.preventScrollReset).toBe(false);
 
       let positions = { "/tasks": 100 };
@@ -5916,7 +6512,6 @@ describe("a router", () => {
         formData: createFormData({}),
       });
       await nav1.actions.tasks.resolve("ACTION");
-      await nav1.loaders.root.resolve("ROOT");
       await nav1.loaders.tasks.resolve("TASKS");
       expect(t.router.state.restoreScrollPosition).toBe(false);
       expect(t.router.state.preventScrollReset).toBe(false);
@@ -5924,17 +6519,16 @@ describe("a router", () => {
 
     it("does not reset scroll", async () => {
       let t = setup({
-        routes: TASK_ROUTES,
+        routes: SCROLL_ROUTES,
         initialEntries: ["/"],
         hydrationData: {
           loaderData: {
-            root: "ROOT_DATA",
             index: "INDEX_DATA",
           },
         },
       });
 
-      expect(t.router.state.restoreScrollPosition).toBe(null);
+      expect(t.router.state.restoreScrollPosition).toBe(false);
       expect(t.router.state.preventScrollReset).toBe(false);
 
       let positions = {};
@@ -6165,8 +6759,8 @@ describe("a router", () => {
             pathname: "/tasks",
             search: "?key=value",
           },
-          formMethod: undefined,
-          formData: undefined,
+          formMethod: "get",
+          formData: createFormData({ key: "value" }),
         },
         revalidation: "loading",
         loaderData: {
@@ -6260,7 +6854,7 @@ describe("a router", () => {
       await N.loaders.root.resolve("ROOT_DATA*");
       await N.loaders.tasks.resolve("TASKS_DATA");
       expect(t.router.state).toMatchObject({
-        historyAction: "REPLACE",
+        historyAction: "PUSH",
         location: { pathname: "/tasks" },
         navigation: IDLE_NAVIGATION,
         revalidation: "idle",
@@ -6272,7 +6866,7 @@ describe("a router", () => {
           tasks: "TASKS_ACTION",
         },
       });
-      expect(t.history.replace).toHaveBeenCalledWith(
+      expect(t.history.push).toHaveBeenCalledWith(
         t.router.state.location,
         t.router.state.location.state
       );
@@ -6472,7 +7066,7 @@ describe("a router", () => {
       await R.loaders.root.resolve("ROOT_DATA*");
       await R.loaders.tasks.resolve("TASKS_DATA*");
       expect(t.router.state).toMatchObject({
-        historyAction: "REPLACE",
+        historyAction: "PUSH",
         location: { pathname: "/tasks" },
         navigation: IDLE_NAVIGATION,
         revalidation: "idle",
@@ -6481,7 +7075,7 @@ describe("a router", () => {
           tasks: "TASKS_DATA*",
         },
       });
-      expect(t.history.replace).toHaveBeenCalledWith(
+      expect(t.history.push).toHaveBeenCalledWith(
         t.router.state.location,
         t.router.state.location.state
       );
@@ -6559,7 +7153,7 @@ describe("a router", () => {
       await R.loaders.root.resolve("ROOT_DATA*");
       await R.loaders.tasks.resolve("TASKS_DATA*");
       expect(t.router.state).toMatchObject({
-        historyAction: "REPLACE",
+        historyAction: "PUSH",
         location: { pathname: "/tasks" },
         navigation: IDLE_NAVIGATION,
         revalidation: "idle",
@@ -6571,7 +7165,7 @@ describe("a router", () => {
           tasks: "TASKS_ACTION",
         },
       });
-      expect(t.history.replace).toHaveBeenCalledWith(
+      expect(t.history.push).toHaveBeenCalledWith(
         t.router.state.location,
         t.router.state.location.state
       );
@@ -6690,7 +7284,7 @@ describe("a router", () => {
         navigation: IDLE_NAVIGATION,
         revalidation: "idle",
         loaderData: {
-          root: "ROOT_DATA",
+          root: undefined,
           index: "INDEX_DATA*",
         },
         errors: {
@@ -6880,6 +7474,7 @@ describe("a router", () => {
           formEncType: undefined,
           formData: undefined,
           data: undefined,
+          " _hasFetcherDoneAnything ": true,
         });
 
         await dfd.resolve("DATA");
@@ -6889,6 +7484,7 @@ describe("a router", () => {
           formEncType: undefined,
           formData: undefined,
           data: "DATA",
+          " _hasFetcherDoneAnything ": true,
         });
 
         expect(router._internalFetchControllers.size).toBe(0);
@@ -6933,6 +7529,10 @@ describe("a router", () => {
           formData: createFormData({ key: "value" }),
         });
         expect(A.fetcher.state).toBe("loading");
+        expect(A.fetcher.formMethod).toBe("get");
+        expect(A.fetcher.formAction).toBe("/foo");
+        expect(A.fetcher.formData).toEqual(createFormData({ key: "value" }));
+        expect(A.fetcher.formEncType).toBe("application/x-www-form-urlencoded");
         expect(
           new URL(
             A.loaders.foo.stub.mock.calls[0][0].request.url
@@ -7377,6 +7977,7 @@ describe("a router", () => {
           formAction: undefined,
           formEncType: undefined,
           formData: undefined,
+          " _hasFetcherDoneAnything ": true,
         });
         expect(t.router.state.historyAction).toBe("PUSH");
         expect(t.router.state.location.pathname).toBe("/bar");
@@ -10446,6 +11047,57 @@ describe("a router", () => {
         });
       });
 
+      it("should fill in null loaderData values for routes without loaders", async () => {
+        let { query } = createStaticHandler([
+          {
+            id: "root",
+            path: "/",
+            children: [
+              {
+                id: "none",
+                path: "none",
+              },
+              {
+                id: "a",
+                path: "a",
+                loader: () => "A",
+                children: [
+                  {
+                    id: "b",
+                    path: "b",
+                  },
+                ],
+              },
+            ],
+          },
+        ]);
+
+        // No loaders at all
+        let context = await query(createRequest("/none"));
+        expect(context).toMatchObject({
+          actionData: null,
+          loaderData: {
+            root: null,
+            none: null,
+          },
+          errors: null,
+          location: { pathname: "/none" },
+        });
+
+        // Mix of loaders and no loaders
+        context = await query(createRequest("/a/b"));
+        expect(context).toMatchObject({
+          actionData: null,
+          loaderData: {
+            root: null,
+            a: "A",
+            b: null,
+          },
+          errors: null,
+          location: { pathname: "/a/b" },
+        });
+      });
+
       it("should support document load navigations returning responses", async () => {
         let { query } = createStaticHandler(SSR_ROUTES);
         let context = await query(createRequest("/parent/json"));
@@ -10750,6 +11402,61 @@ describe("a router", () => {
         });
       });
 
+      it("should handle multiple errors at separate boundaries", async () => {
+        let routes = [
+          {
+            id: "root",
+            path: "/",
+            loader: () => Promise.reject("ROOT"),
+            hasErrorBoundary: true,
+            children: [
+              {
+                id: "child",
+                path: "child",
+                loader: () => Promise.reject("CHILD"),
+                hasErrorBoundary: true,
+              },
+            ],
+          },
+        ];
+
+        let { query } = createStaticHandler(routes);
+        let context;
+
+        context = await query(createRequest("/child"));
+        expect(context.errors).toEqual({
+          root: "ROOT",
+          child: "CHILD",
+        });
+      });
+
+      it("should handle multiple errors at the same boundary", async () => {
+        let routes = [
+          {
+            id: "root",
+            path: "/",
+            loader: () => Promise.reject("ROOT"),
+            hasErrorBoundary: true,
+            children: [
+              {
+                id: "child",
+                path: "child",
+                loader: () => Promise.reject("CHILD"),
+              },
+            ],
+          },
+        ];
+
+        let { query } = createStaticHandler(routes);
+        let context;
+
+        context = await query(createRequest("/child"));
+        expect(context.errors).toEqual({
+          // higher error value wins
+          root: "ROOT",
+        });
+      });
+
       it("should handle aborted load requests", async () => {
         let dfd = createDeferred();
         let controller = new AbortController();
@@ -10911,7 +11618,13 @@ describe("a router", () => {
             ],
           },
         ]);
-        await query(createSubmitRequest("/child"));
+        await query(
+          createSubmitRequest("/child", {
+            headers: {
+              test: "value",
+            },
+          })
+        );
 
         // @ts-expect-error
         let actionRequest = actionStub.mock.calls[0][0]?.request;
@@ -10928,8 +11641,12 @@ describe("a router", () => {
         let childLoaderRequest = childLoaderStub.mock.calls[0][0]?.request;
         expect(rootLoaderRequest.method).toBe("GET");
         expect(rootLoaderRequest.url).toBe("http://localhost/child");
+        expect(rootLoaderRequest.headers.get("test")).toBe("value");
+        expect(await rootLoaderRequest.text()).toBe("");
         expect(childLoaderRequest.method).toBe("GET");
         expect(childLoaderRequest.url).toBe("http://localhost/child");
+        expect(childLoaderRequest.headers.get("test")).toBe("value");
+        // Can't re-read body here since it's the same request as the root
       });
 
       it("should support a requestContext passed to loaders and actions", async () => {